Rename setupfn to SetupFormComponent
