+++ src/components/BaseRadio.vue
@@ -10,7 +10,7 @@
 </template>
 
 <script>
-import SetupFormInput from '@/features/SetupFormInput'
+import SetupFormComponent from '@/features/SetupFormComponent'
 
 export default {
   inheritAttrs: false,
@@ -27,7 +27,7 @@ export default {
     }
   },
   setup (props, context) {
-    const { updateValue } = SetupFormInput(props, context)
+    const { updateValue } = SetupFormComponent(props, context)
 
     return {
       updateValue
