+++ src/components/BaseSelect.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script>
-import SetupFormInput from '@/features/SetupFormInput'
+import SetupFormComponent from '@/features/SetupFormComponent'
 
 export default {
   inheritAttrs: false,
@@ -36,7 +36,7 @@ export default {
     }
   },
   setup (props, context) {
-    const { updateValue } = SetupFormInput(props, context)
+    const { updateValue } = SetupFormComponent(props, context)
 
     return {
       updateValue
